observability-control-plane/requirements.txt
memori>=0.3.1
httpx>=0.27.0
flask>=3.0.0
flask-htmx>=0.3.0
py2neo>=2021.2.0
psycopg2-binary>=2.9.0
dagger-io>=0.12.0
psutil>=5.9.0
docker>=7.0.0
